Add useMemory hook
